Factor tooltip for existing feedback out of `SelectionRenderer`
